requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.2.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.8.2
bleach==3.3.0
boto3==1.13.2
botocore==1.16.2
branca==0.3.1
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.8
django-cleanup==4.0.0
django-crispy-forms==1.8.1
django-heroku==0.3.1
django-markdownx==3.0.1
django-storages==1.9.1
docutils==0.15.2
entrypoints==0.3
folium==0.10.1
gunicorn==20.0.4
idna==2.8
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.2
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
notebook==6.2.0
numpy==1.18.4
packaging==20.9
panda==0.3.1
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
powerline-status==2.7
prometheus-client==0.9.0
prompt-toolkit==3.0.16
psycopg2==2.8.5
psycopg2-binary==2.8.5
ptyprocess==0.7.0
public==2019.4.13
pycparser==2.20
pygame==2.0.0.dev6
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
query-string==2019.4.13
requests==2.22.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
statistics==1.0.3.5
terminado==0.9.2
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==0.16.1
whitenoise==5.0.1
widgetsnbextension==3.5.1
